areglos nombres de resquesty organizacion de controladores por rol

+++ routes/web.php
@@ -13,40 +13,40 @@
 
  
 Auth::routes();
-Route::get('/', 'BannerController@show')->name('landing');
+Route::get('/', 'Company\BannerController@show')->name('landing');
 Route::get('/home', 'HomeController@index')->name('home');//home general
 Route::view('/login', 'auth.login')->name('login');
 Route::post('/login', 'Auth\LoginController@login')->name("login_p");
 Route::get('/logout', 'Auth\LoginController@logout')->name("logout");
 Route::view('/registro', 'auth.register')->name('registro');
 
-
+ 
 Route::group(['middleware'=>['check.admin.role']], function(){
 //RUTAS ADMIN
 //Route::view('/admin/plans', 'admin/plans')->name('admin plans');
-Route::get('/admin/users', 'UsersController@index')->name('users');
+Route::get('/admin/users', 'Admin\UserController@index')->name('users');
 
 //Rutas CRUD de planes
-Route::get('/admin/plans', 'PlanController@index')->name('plans');
-Route::get('/admin/plans/add', 'PlanController@create')->name('add plan');
-Route::post('/admin/plans/create', 'PlanController@store')->name('create plan');
-Route::delete('admin/plans/delete/{id}', 'PlanController@destroy')->name('destroy plan');
+Route::get('/admin/plans', 'Admin\PlanController@index')->name('plans');
+Route::get('/admin/plans/add', 'Admin\PlanController@create')->name('add plan');
+Route::post('/admin/plans/create', 'Admin\PlanController@store')->name('create plan');
+Route::delete('admin/plans/delete/{id}', 'Admin\PlanController@destroy')->name('destroy plan');
 
 //Rutas CRUD de RouteTypes
-Route::get('/admin/routeTypes', 'RouteTypeController@index')->name('route Types');
-Route::get('/admin/routeTypes/add', 'RouteTypeController@create')->name('add route Types');
-Route::post('/admin/routeTypes/create', 'RouteTypeController@store')->name('create route Types');
-Route::Delete('/admin/routeTypes/delete/{id}', 'RouteTypeController@destroy')->name('destroy route Types');
+Route::get('/admin/routeTypes', 'Admin\RouteTypeController@index')->name('route Types');
+Route::get('/admin/routeTypes/add', 'Admin\RouteTypeController@create')->name('add route Types');
+Route::post('/admin/routeTypes/create', 'Admin\RouteTypeController@store')->name('create route Types');
+Route::Delete('/admin/routeTypes/delete/{id}', 'Admin\RouteTypeController@destroy')->name('destroy route Types');
 });
 
 Route::group(['middleware'=>['check.user.role']], function(){
 //RUTAS USER
-Route::get('/user/rutas', 'RouteController@index')->name('user routes');
-Route::get('/user/ruta/{id}', 'RouteController@show')->name('user show route');
-Route::get('/user/editar/ruta/{id}','RouteController@edit')->name('user edit route');
-Route::post('/user/guardar/ruta','RouteController@store')->name('user store routes');
-Route::put('/user/actualizar/ruta/{id}', 'RouteController@update')->name('user update route');
-Route::delete('/user/borrar/ruta/{id}','RouteController@destroy')->name('user destroy route');
+Route::get('/user/rutas', 'User\RouteController@index')->name('user routes');
+Route::get('/user/ruta/{id}', 'User\RouteController@show')->name('user show route');
+Route::get('/user/editar/ruta/{id}','User\RouteController@edit')->name('user edit route');
+Route::post('/user/guardar/ruta','User\RouteController@store')->name('user store routes');
+Route::put('/user/actualizar/ruta/{id}', 'User\RouteController@update')->name('user update route');
+Route::delete('/user/borrar/ruta/{id}','User\RouteController@destroy')->name('user destroy route');
 Route::view('/user/opciones', 'user/options')->name('user options');
 
 
@@ -56,9 +56,9 @@ Route::group(['middleware'=>['check.company.role']], function(){
 //RUTAS COMPANY
 Route::view('/company/plan', 'company/planes')->name('company plan');
 Route::view('/company/location', 'company/location')->name('company location');
-Route::get('/company/banner', 'BannerController@index')->name('company banner');
-Route::get('/company/banner/add', 'BannerController@create')->name('company banner add');
-Route::post('/company/banner/save', 'BannerController@store')->name('company banner store');
+Route::get('/company/banner', 'Company\BannerController@index')->name('company banner');
+Route::get('/company/banner/add', 'Company\BannerController@create')->name('company banner add');
+Route::post('/company/banner/save', 'Company\BannerController@store')->name('company banner store');
 
 });
